requirements.txt
pandas
requests
python-dotenv
numpy
pyyaml
yfinance
yfinance